+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -178,3 +178,4 @@ CHANGELOG
  * made the normalized data of a form accessible in the template through the variable "form.vars.data"
  * made the original data of a choice accessible in the template through the property "choice.data"
  * added convenience class Forms and FormFactoryBuilderInterface
+ * buildForm is now abstract in AbstractType